requirements.txt
uvicorn==0.17.5
fastapi~=0.78.0
pandas==1.3.5
numpy==1.21.6
python-multipart==0.0.5
mongoengine==0.29.1
passlib==1.7.4
python-jose==3.3.0
python-dotenv
pymongo==4.9
bcrypt
